Rename tests to a new standard
