+++ galasa-inttests-parent/dev.galasa.inttests/src/main/java/dev/galasa/inttests/simbank/kubernetes/SimBankKubernetesJava08.java
@@ -1,9 +1,9 @@
 /*
  * Licensed Materials - Property of IBM
  * 
- * (c) Copyright IBM Corp. 2020.
+ * (c) Copyright IBM Corp. 2020,2021.
  */
-package dev.galasa.inttests;
+package dev.galasa.inttests.simbank.kubernetes;
 
 import java.util.UUID;
 
@@ -13,19 +13,18 @@ import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 
-import dev.galasa.Summary;
 import dev.galasa.Test;
 import dev.galasa.core.manager.Logger;
 import dev.galasa.core.manager.RunName;
 import dev.galasa.galasaecosystem.GalasaEcosystemManagerException;
 import dev.galasa.galasaecosystem.IKubernetesEcosystem;
 import dev.galasa.galasaecosystem.KubernetesEcosystem;
+import dev.galasa.inttests.TestException;
 import dev.galasa.kubernetes.IKubernetesNamespace;
 import dev.galasa.kubernetes.KubernetesNamespace;
 
 @Test
-@Summary("Run Artifact IVT tests in a provisioned Galasa Ecosystem")
-public class ArtifactIntegrationTests {
+public class SimBankKubernetesJava08 {
 
     @Logger
     public Log logger;
@@ -41,12 +40,38 @@ public class ArtifactIntegrationTests {
     
     
     @Test
-    public void testArtifactManagerIVT() throws GalasaEcosystemManagerException, InterruptedException, TestException {
+    public void testSimbankIvt() throws GalasaEcosystemManagerException, InterruptedException, TestException {
 
-        submitTest("dev.galasa.artifact.manager.ivt.ArtifactManagerIVT");
+        submitTest("dev.galasa.simbank.tests.SimBankIVT");
         
-        logger.info("ArtifactManagerIVT test passed");
+        logger.info("SimBankIVT test passed");
     }
+    
+    @Test
+    public void testBasicAccountCreditTest() throws GalasaEcosystemManagerException, InterruptedException, TestException {
+
+        submitTest("dev.galasa.simbank.tests.BasicAccountCreditTest");
+        
+        logger.info("BasicAccountCreditTest test passed");
+    }
+    
+    @Test
+    public void testProvisionedAccountCreditTests() throws GalasaEcosystemManagerException, InterruptedException, TestException {
+
+        submitTest("dev.galasa.simbank.tests.ProvisionedAccountCreditTests");
+        
+        logger.info("ProvisionedAccountCreditTests test passed");
+    }
+    
+//    @Test
+//    public void testBatchAccountsOpenTest() throws GalasaEcosystemManagerException, InterruptedException, TestException {
+//
+//        submitTest("dev.galasa.simbank.tests.BatchAccountsOpenTest");
+//        
+//        logger.info("BatchAccountsOpenTest test passed");
+//    }
+    
+    
     
     
     private void submitTest(String testName) throws GalasaEcosystemManagerException, TestException {
@@ -57,11 +82,11 @@ public class ArtifactIntegrationTests {
         ecosystem.submitRun(null, 
                 runName,
                 groupName, 
-                "dev.galasa.artifact.manager.ivt", 
+                "dev.galasa.simbank.tests", 
                 testName, 
                 null, 
                 null, 
-                null, 
+                "simbank", 
                 null);
         
         JsonObject finalResponse = ecosystem.waitForGroupNames(groupName, 180);
@@ -84,7 +109,5 @@ public class ArtifactIntegrationTests {
             throw new TestException("Run did not pass, last response was:-\n" + finalResponse);
         }   
     }
-    
-
     
 }
